==> R_scripts/02a_DataExploration.Rmd <==
---
title: "R Notebook"
output: html_notebook
---

Data exploration

```{r}
library(microbiome)
```

# Read-in Data

```{r Read in data, warning=FALSE, echo=FALSE}
phylo.spring <- readRDS(file = "../Data/phylo-spring.rds")
metadata <- meta(phylo.spring) # doesnt maintain Date data-type
```

Covariates: ageBinned * habitat * layDateFirst + ageBinned * broodSizeWhenSampled + habitat*DistanceToEdge +  (1|site/nest/bird.ID) + (1|SequencePlate)
- ageBinned * habitat * laydateFirst
- ageBinned * broodSizeWhensampled
- habitat*DistanceToEdge
- (1|site/nest/bird.ID)
- (1|SequencePlate)

Does each covariate have adequate number of samples?
* many more deciduous than coniferous samples (particularly adults)
* samples evenly distributed across laydate by age
* samples relatively evenly distributed in terms of brood size across ages
* samples distance to edge evenly distributed across habitats
```{r}
table(metadata$ageBinned, metadata$habitat)

plot(metadata$ageBinned, metadata$layDateFirst)

table(metadata$ageBinned, metadata$broodSizeWhenSampled)
plot(metadata$ageBinned, metadata$broodSizeWhenSampled)

plot(metadata$habitat, metadata$DistanceToEdge)
```


Are predictors colinear
* some pattern apparent in distanceToedge vs layDate
```{r}
df <- metadata[,c("ageBinned", "habitat", "broodSizeWhenSampled", "layDateFirst", "DistanceToEdge")]
pairs(df)
```

Positive correlations in red, negative in blue, intensity if colour indicates strength of correlation
```{r}
#install.packages("corrgram")
library(corrgram)

corrgram(df)
```
