==> R_scripts/04_BetaDiversityAnalysis.Rmd <==
---
title: "Host and Environmental drivers of microbiome: Beta diversity"
output: html_notebook
---

This script carries out the beta diversity analyses (community composition and structure) for the 2016 Cork great tit microbiome data.

Overall plan"
* model 1: all variables of interest, nest as blocking factor
* model 2: all variables of interest but age-habitat interaction, nest blocking
* model 3: adult samples, all variables of interest, nest blocking
* then check are results different when individuals w/ 2 samples are dropped

Gabrielle says not to remove individuals sampled twice

# Set-up and libraries

```{r Housekeeping, echo=FALSE}
rm(list=ls())
#set.seed(10)
```


```{r Set-up, message=FALSE, echo=FALSE}
# libraries
#install.packages("tidyverse", dependencies = T)
library(vegan)  ##note had trouble installing, needed to install gfortran first
library(tidyverse) #
library(phyloseq) 
library(microbiome)
library(grid)
library(gridExtra)
library(ggpubr)
library(kableExtra)
library(compositions)
library(ggfortify)
library(ggplot2)
library(cowplot)

library(arm)
```

```{r}
#library(devtools)
#install_github("pmartinezarbizu/pairwiseAdonis/pairwiseAdonis")
library(pairwiseAdonis)
```

```{r Read in data, warning=FALSE, echo=FALSE}
phylo.spring <- readRDS(file = "../Data/phylo-spring.rds")
metadata <- meta(phylo.spring) # doesnt maintain Date data-type

phylo.spring
```

# Data cleaning

```{r Data splits, echo=FALSE}

# relevel habitat
#metadata.NoDups$habitat <- relevel(metadata.NoDups$habitat, "deciduous")

# # subset to adults
phylo.adults <- subset_samples(phylo.spring, ageBinned=="adult")
adults.meta <- meta(phylo.adults@sam_data)

```

Do i need to scale numeric variables?
```{r}
numeric.predictors <- c("QubitDNA","Tarsus","Weight", "wing", "broodSizeWhenSampled", "broodSizeMax", "totalFledge", "clutchSize", "layDateFirst", "numberDeadPreRinged", "numberDeadPostRinged","scaled.mass","scaled.mass.wing.adult", "scaled.mass.tarsus.adult","scaled.mass.chick", "scaled.mass.tarsus", "scaled.mass.wing", "DistanceToEdge")

# centre and scale numeric variables ie. subtract mean and divide by st. deviation
metadata.scaled <- metadata
metadata.scaled[,numeric.predictors] <- scale(metadata.scaled[,numeric.predictors])

adults.meta.scaled <- adults.meta
adults.meta.scaled[,numeric.predictors] <- scale(adults.meta.scaled[,numeric.predictors])
```

# All birds: CoDa, w/ duplicates

Compositional method, use clr transform

Keeping repeated measures here but using blocking factor to control for repeated samples. 

```{r}
# Filter rare taxa
phylo.knowles <- filter_taxa(phylo.spring, function(x) sum(x > 1) > (0.05*length(x)), TRUE)
```

```{r}
# phylo.knowles.clr <- phylo.knowles
# phylo.knowles.clr@otu_table <- otu_table(clr(phylo.knowles@otu_table), taxa_are_rows = F)
# aitchison.dist.dups <- phyloseq::distance(phylo.knowles.clr, method = "euclidean")
pk.otu.clr.dups <- clr(phylo.knowles@otu_table)
aitchison.dist.dups <- vegdist(pk.otu.clr.dups, method = "euclid")
```

```{r}
# phylo.TSS <- transform_sample_counts(phylo.knowles, function(x) x/sum(x)) # normalise read counts w/ Total-Sum Scaling
# 
# BCdist <- phyloseq::distance(phylo.TSS, method="bray")
```

All variables except for habitat have heterogenous dispersions
```{r, results='hide', warning=FALSE}
## H0= No difference in dispersion between groups
# calc dispersion, using distance measure

variables <- c("ageBinned", "habitat", "layDateFirst", "broodSizeWhenSampled", "DistanceToEdge", "SequencePlate")

for(i in variables){ # works
  dispersion <- betadisper(aitchison.dist.dups, metadata.scaled[,i])  #, bias.adjust = T
  print(i) # print variable being tested
  print(permutest(dispersion, pairwise=FALSE, permutations=1000))
  cat("\n") # print line break, makes it easier to read
}
```

```{r}
#hist(metadata$layDateFirst)
#ist(sqrt(metadata$DistanceToEdge))

dispersion1 <- betadisper(aitchison.dist.dups, metadata$ageBinned)  #, bias.adjust = T
print(permutest(dispersion1, pairwise=FALSE, permutations=1000))

dispersion2 <- betadisper(aitchison.dist.dups, metadata$habitat)  #, bias.adjust = T
print(permutest(dispersion2, pairwise=FALSE, permutations=1000))

dispersion3 <- betadisper(aitchison.dist.dups, (metadata$layDateFirst))  #, bias.adjust = T
print(permutest(dispersion3, pairwise=FALSE, permutations=1000))

dispersion4 <- betadisper(aitchison.dist.dups, metadata$broodSizeWhenSampled)  #, bias.adjust = T
print(permutest(dispersion4, pairwise=FALSE, permutations=1000))

dispersion5 <- betadisper(aitchison.dist.dups, (metadata$DistanceToEdge))  #, bias.adjust = T
print(permutest(dispersion5, pairwise=FALSE, permutations=1000))

dispersion6 <- betadisper(aitchison.dist.dups, metadata$SequencePlate)  #, bias.adjust = T
print(permutest(dispersion6, pairwise=FALSE, permutations=1000))
```

```{r}
TukeyHSD(dispersion1)
TukeyHSD(dispersion2)
#TukeyHSD(dispersion3)
#TukeyHSD(dispersion4)
#TukeyHSD(dispersion5)
```

Plot dispersions
```{r}
plot(dispersion1)
plot(dispersion4)
```
Add below chunk to test effect of nest for discussion
```{r}
check.nest <- adonis2(aitchison.dist.dups ~ ageBinned + habitat + layDateFirst + broodSizeWhenSampled + DistanceToEdge + SequencePlate + nest, by="margin", method="euclidian", data = metadata.scaled)

check.nest
```

## Main model
```{r}
perms.dups <- with(metadata, how(nperm = 1000, blocks = nest))

all.adonis.dups.fixed <- adonis2(aitchison.dist.dups ~ ageBinned + habitat + layDateFirst + broodSizeWhenSampled + DistanceToEdge + SequencePlate, by="margin", method="euclidian", data = metadata.scaled, permutations = perms.dups)

all.adonis.dups.fixed
```


## Pairwise adonis, duplicates
```{r}
pairwiseAdonis::pairwise.adonis2(aitchison.dist.dups ~ ageBinned * habitat, data = metadata.scaled)
#?pairwise.adonis2()
```

Try same as above but strata by sequence plate and then include individual ID
```{r}
# perms.dups <- with(metadata, how(nperm = 1000, blocks = SequencePlate))
# 
# all.adonis.dups.fixed <- adonis2(aitchison.dist.dups ~ ageBinned + habitat + layDateFirst + broodSizeWhenSampled + DistanceToEdge + bird.ID, by="margin", method="euclidian", data = metadata.scaled, permutations = perms.dups)
# 
# all.adonis.dups.fixed
```

Including bird.ID as a fixed effect and blocking by nest suggests age is in fact a significant factor though accounts for only 0.7% of variation, while bird id accounts for 75%, but is non-significant. Sequence plate is significant, accounting for 2.5%, brood size is not significant while habitat, lay date and distance to edge could not be estimated.
```{r}
# perms.dups <- with(metadata, how(nperm = 1000, blocks = nest))
# 
# all.adonis.dups.fixed <- adonis2(aitchison.dist.dups ~ ageBinned + habitat + layDateFirst + broodSizeWhenSampled + DistanceToEdge + SequencePlate + bird.ID, by="margin", method="euclidian", data = metadata.scaled, permutations = perms.dups)
# 
# all.adonis.dups.fixed
```

```{r}
# perms.dups <- with(metadata, how(nperm = 1000, blocks = bird.ID))
# 
# all.adonis.dups.fixed <- adonis2(aitchison.dist.dups ~ ageBinned + habitat + layDateFirst + broodSizeWhenSampled + DistanceToEdge + SequencePlate, by="margin", method="euclidian", data = metadata.scaled, permutations = perms.dups)
# 
# all.adonis.dups.fixed
```

## Model, interaction age x habitat
```{r}
perms.dups <- with(metadata, how(nperm = 1000, blocks = nest))

all.adonis.dups.int <- adonis2(aitchison.dist.dups ~ ageBinned*habitat + layDateFirst + broodSizeWhenSampled + DistanceToEdge + SequencePlate, by="margin", data = metadata.scaled, permutations = perms.dups) #
all.adonis.dups.int
```

Interaction calc by term rather than margin to get estimates for main effects as well as interaction term
```{r}
perms.dups <- with(metadata, how(nperm = 1000, blocks = nest))

adonis2(aitchison.dist.dups ~ ageBinned*habitat + layDateFirst + broodSizeWhenSampled + DistanceToEdge + SequencePlate, by="term", data = metadata.scaled, permutations = perms.dups) #

```

## check influence of unbalanced design
 *see 04a___balanced script*

# All birds, w/o duplicates
Reviewer is asking about pseudoreplication due to repeated samples- does it influence results

## create no duplicate data

Remove individuals measured twice
* set.seed so same individuals dropped
* no adults measured twice
```{r}
#n_occur <- data.frame(table(metadata.NoDups$bird.ID))
#n_occur[n_occur$Freq > 1,]

set.seed(1189)
metadata.NoDups <- metadata %>% 
    group_by(bird.ID) %>%
    sample_n(1)

dropped.samples <- setdiff(metadata$BIOM.ID, metadata.NoDups$BIOM.ID)

# metadata.even <- metadata %>% 
#     group_by(bird.ID) %>%
#     sample_n(min(table(metadata$bird.ID)))

table(metadata$ageBinned)
table(metadata.NoDups$ageBinned)
```

```{r}
#remove dropped samples
phylo.NoDuplicates <- subset_samples(phylo.spring, BIOM.ID %in%  metadata.NoDups$BIOM.ID)

setdiff(meta(phylo.NoDuplicates)$BIOM.ID , metadata.NoDups$BIOM.ID)
setdiff( metadata.NoDups$BIOM.ID, meta(phylo.NoDuplicates)$BIOM.ID)
```

```{r}
# centre and scale numeric variables ie. subtract mean and divide by st. deviation
metadata.NoDups.scaled <- metadata.NoDups
metadata.NoDups.scaled[,numeric.predictors] <- scale(metadata.NoDups.scaled[,numeric.predictors])
```

## filter and transform
```{r}
# Filter rare taxa
phylo.knowles.NoDuplicates <- filter_taxa(phylo.NoDuplicates, function(x) sum(x > 1) > (0.05*length(x)), TRUE)
```

```{r}
pk.otu.clr.Nodups <- clr(phylo.knowles.NoDuplicates@otu_table)
aitchison.dist.Nodups <- vegdist(pk.otu.clr.Nodups, method = "euclid")
```

## model
```{r}
perms.Nodups <- with(metadata.NoDups.scaled, how(nperm = 1000, blocks = nest))
# 
all.adonis.Nodups.fixed <- adonis2(aitchison.dist.Nodups ~ ageBinned + habitat + layDateFirst + broodSizeWhenSampled + DistanceToEdge + SequencePlate, by="margin", method="euclidian", data = metadata.NoDups.scaled, permutations = perms.Nodups)
# 
all.adonis.Nodups.fixed
```

```{r}
perms.Nodups <- with(metadata.NoDups.scaled, how(nperm = 1000, blocks = nest))
# 
all.adonis.Nodups.int <- adonis2(aitchison.dist.Nodups ~ ageBinned*habitat + layDateFirst + broodSizeWhenSampled + DistanceToEdge + SequencePlate, by="margin", data = metadata.NoDups.scaled, permutations = perms.Nodups) #

all.adonis.Nodups.int
```

Interaction calc by term rather than margin to get estimates for main effects as well as interaction term
```{r}
# perms.dups <- with(metadata, how(nperm = 1000, blocks = nest))
# 
# adonis2(aitchison.dist.dups ~ ageBinned*habitat + layDateFirst + broodSizeWhenSampled + DistanceToEdge + SequencePlate, by="term", data = metadata.scaled, permutations = perms.dups) #

```

## test dispersion for no duplicate dataset
```{r}
#hist(metadata$layDateFirst)
#ist(sqrt(metadata$DistanceToEdge))

dispersion1 <- betadisper(aitchison.dist.Nodups, metadata.NoDups$ageBinned)  #, bias.adjust = T
print(permutest(dispersion1, pairwise=FALSE, permutations=1000))

dispersion2 <- betadisper(aitchison.dist.Nodups, metadata.NoDups$habitat)  #, bias.adjust = T
print(permutest(dispersion2, pairwise=FALSE, permutations=1000))

dispersion3 <- betadisper(aitchison.dist.Nodups, (metadata.NoDups$layDateFirst))  #, bias.adjust = T
print(permutest(dispersion3, pairwise=FALSE, permutations=1000))

dispersion4 <- betadisper(aitchison.dist.Nodups, metadata.NoDups$broodSizeWhenSampled)  #, bias.adjust = T
print(permutest(dispersion4, pairwise=FALSE, permutations=1000))

dispersion5 <- betadisper(aitchison.dist.Nodups, (metadata.NoDups$DistanceToEdge))  #, bias.adjust = T
print(permutest(dispersion5, pairwise=FALSE, permutations=1000))

dispersion6 <- betadisper(aitchison.dist.Nodups, metadata.NoDups$SequencePlate)  #, bias.adjust = T
print(permutest(dispersion6, pairwise=FALSE, permutations=1000))
```

## Pairwise adonis, no duplicates
```{r}
pairwiseAdonis::pairwise.adonis2(aitchison.dist.Nodups ~ ageBinned + habitat, data = metadata.NoDups.scaled)
#?pairwise.adonis2()
```

# D8+adults
```{r}
phylo.d8 <- subset_samples(phylo.spring, ageBinned!="2week")
d8.meta <- meta(phylo.d8@sam_data)


d8.meta.scaled <- d8.meta
d8.meta.scaled[,numeric.predictors] <- scale(d8.meta.scaled[,numeric.predictors])
```

```{r}
# Filter rare taxa
phylo.knowles.d8 <- filter_taxa(phylo.d8, function(x) sum(x > 1) > (0.05*length(x)), TRUE)
```

```{r}
d8.clr <- clr(phylo.d8@otu_table)
aitch.d8 <- vegdist(d8.clr, method = "euclid")
```

## Main model
```{r}
perms.dups <- with(d8.meta, how(nperm = 1000, blocks = nest))

d8.adonis <- adonis2(aitch.d8 ~ ageBinned * habitat + layDateFirst + broodSizeWhenSampled + DistanceToEdge + SequencePlate, by="margin", method="euclidian", data = d8.meta.scaled, permutations = perms.dups)

d8.adonis
```

## Main model, fixed only
```{r}
perms.dups <- with(d8.meta, how(nperm = 1000, blocks = nest))

d8.adonis <- adonis2(aitch.d8 ~ ageBinned + habitat + layDateFirst + broodSizeWhenSampled + DistanceToEdge + SequencePlate, by="margin", method="euclidian", data = d8.meta.scaled, permutations = perms.dups)

d8.adonis
```

### Nest as fixed
```{r}
#perms.dups <- with(d8.meta, how(nperm = 1000, blocks = nest))

d8.adonis <- adonis2(aitch.d8 ~ ageBinned + habitat + layDateFirst + broodSizeWhenSampled + DistanceToEdge + SequencePlate + nest, by="margin", method="euclidian", data = d8.meta.scaled)

d8.adonis
```

## test dispersion for D8-adult dataset
```{r}
#hist(metadata$layDateFirst)
#ist(sqrt(metadata$DistanceToEdge))

dispersion1 <- betadisper(aitch.d8, d8.meta$ageBinned)  #, bias.adjust = T
print(permutest(dispersion1, pairwise=FALSE, permutations=1000))

dispersion2 <- betadisper(aitch.d8, d8.meta$habitat)  #, bias.adjust = T
print(permutest(dispersion2, pairwise=FALSE, permutations=1000))

dispersion3 <- betadisper(aitch.d8, d8.meta$layDateFirst)  #, bias.adjust = T
print(permutest(dispersion3, pairwise=FALSE, permutations=1000))

dispersion4 <- betadisper(aitch.d8, d8.meta$broodSizeWhenSampled)  #, bias.adjust = T
print(permutest(dispersion4, pairwise=FALSE, permutations=1000))

dispersion5 <- betadisper(aitch.d8, d8.meta$DistanceToEdge)  #, bias.adjust = T
print(permutest(dispersion5, pairwise=FALSE, permutations=1000))

dispersion6 <- betadisper(aitch.d8, d8.meta$SequencePlate)  #, bias.adjust = T
print(permutest(dispersion6, pairwise=FALSE, permutations=1000))
```

# D15+adults
```{r}
phylo.d15 <- subset_samples(phylo.spring, ageBinned!="1week")
d15.meta <- meta(phylo.d15@sam_data)


d15.meta.scaled <- d15.meta
d15.meta.scaled[,numeric.predictors] <- scale(d15.meta.scaled[,numeric.predictors])
```

```{r}
# Filter rare taxa
phylo.knowles.d15 <- filter_taxa(phylo.d15, function(x) sum(x > 1) > (0.05*length(x)), TRUE)
```

```{r}
d15.clr <- clr(phylo.d15@otu_table)
aitch.d15 <- vegdist(d15.clr, method = "euclid")
```

## Main model
```{r}
perms.dups <- with(d15.meta, how(nperm = 1000, blocks = nest))

d15.adonis <- adonis2(aitch.d15 ~ ageBinned * habitat + layDateFirst + broodSizeWhenSampled + DistanceToEdge + SequencePlate, by="margin", method="euclidian", data = d15.meta.scaled, permutations = perms.dups)

d15.adonis
```

## Main model, fixed only
```{r}
perms.dups <- with(d15.meta, how(nperm = 1000, blocks = nest))

d15.adonis <- adonis2(aitch.d15 ~ ageBinned + habitat + layDateFirst + broodSizeWhenSampled + DistanceToEdge + SequencePlate, by="margin", method="euclidian", data = d15.meta.scaled, permutations = perms.dups)

d15.adonis
```

### Nest as fixed
```{r}
#perms.dups <- with(d8.meta, how(nperm = 1000, blocks = nest))

d15.adonis <- adonis2(aitch.d15 ~ ageBinned + habitat + layDateFirst + broodSizeWhenSampled + DistanceToEdge + SequencePlate + nest, by="margin", method="euclidian", data = d15.meta.scaled)

d15.adonis
```

## test dispersion for D15-adult dataset
```{r}
#hist(metadata$layDateFirst)
#ist(sqrt(metadata$DistanceToEdge))

dispersion1 <- betadisper(aitch.d15, d15.meta$ageBinned)  #, bias.adjust = T
print(permutest(dispersion1, pairwise=FALSE, permutations=1000))

dispersion2 <- betadisper(aitch.d15, d15.meta$habitat)  #, bias.adjust = T
print(permutest(dispersion2, pairwise=FALSE, permutations=1000))

dispersion3 <- betadisper(aitch.d15, d15.meta$layDateFirst)  #, bias.adjust = T
print(permutest(dispersion3, pairwise=FALSE, permutations=1000))

dispersion4 <- betadisper(aitch.d15, d15.meta$broodSizeWhenSampled)  #, bias.adjust = T
print(permutest(dispersion4, pairwise=FALSE, permutations=1000))

dispersion5 <- betadisper(aitch.d15, d15.meta$DistanceToEdge)  #, bias.adjust = T
print(permutest(dispersion5, pairwise=FALSE, permutations=1000))

dispersion6 <- betadisper(aitch.d15, d15.meta$SequencePlate)  #, bias.adjust = T
print(permutest(dispersion6, pairwise=FALSE, permutations=1000))
```
# Adult birds: CoDa

Aitchison distance is supposedly robust to subsetting- can i therefore just subset the adult birds from previous distance calculation?

```{r}
phylo.kn.adult <- filter_taxa(phylo.adults, function(x) sum(x > 1) > (0.05*length(x)), TRUE)

# phylo.TSS.adults <- transform_sample_counts(phylo.kn.adult, function(x) x/sum(x))
# 
# BCdist.adults <- phyloseq::distance(phylo.TSS.adults, method="bray")
# JDdist.adults <- phyloseq::distance(phylo.TSS.adults, method="jaccard")
```

```{r}
pk.otu.clr.adult <- clr(phylo.kn.adult@otu_table)
aitchison.dist.adult <- vegdist(pk.otu.clr.adult, method = "euclidean")
```

## Check dispersion

Homogenous dispersion: sex, habitat, distance is almost non-homo
Non-homogenous: ageDays, ...all others
```{r, results='hide', warning=FALSE}
## H0= No difference in dispersion between groups
# calc dispersion, using distance measure

variables <- c("ageDays", "Sex", "habitat", "layDateFirst", "broodSizeWhenSampled", "DistanceToEdge", "SequencePlate")

for(i in variables){ # works
  dispersion <- betadisper(aitchison.dist.adult, adults.meta.scaled[,i])
  print(i) # print variable being tested
  print(permutest(dispersion, pairwise=FALSE, permutations=1000))
  cat("\n") # print line break, makes it easier to read
}

```

## Modelling: adult aitchison

Sex * habitat * broodSizeWhenSampled + Sex * habitat * layDateFirst + habitat * DistanceToEdge + (1|nest) + (1|SequencePlate)
```{r}
perms.adult <- with(adults.meta, how(nperm = 1000, blocks = nest))

adult.adonis <- adonis2(aitchison.dist.adult ~ ageDays + Sex + habitat + DistanceToEdge + layDateFirst + broodSizeWhenSampled + SequencePlate, by = "margin", data = adults.meta.scaled, permutations = perms.adult)
adult.adonis

#adonis2(JDdist.adults ~ Sex + habitat + DistanceToEdge + layDateFirst, by = "margin", data = adults.meta, permutations = perms.adult)
```

# Results

Table of all results for fixed and interaction PERMANOVA plus adult 
```{r}
#make df of lmer output
all.bdiv <- as_tibble(rbind(all.adonis.dups.fixed, all.adonis.dups.int, adult.adonis), rownames="Independent variables") %>%
  dplyr::rename("P_estimate"="Pr(>F)") %>% 
  mutate_if(is.numeric, round, 3) %>%
  mutate(P_estimate=ifelse(P_estimate==0,"<0.001",P_estimate)) %>% 
  mutate(P_estimate=ifelse(P_estimate<=0.05,str_c(P_estimate," *"),P_estimate)) %>%
  mutate(P_estimate=ifelse(P_estimate>=0.05 & P_estimate<=0.06,str_c(P_estimate,"  ."),P_estimate))


# make df into kable
## this is in html, dosnt render in word doc # but can copy-paste
all.bdiv.results <- kable(all.bdiv, format = "html", table.attr = "style = \"color: black;\"") %>%
  kableExtra::kable_styling(full_width = F) #%>%
 # kableExtra::group_rows("(a) Top model, proteobacteria",1,nrow(proteobacteria.modelAvg.df)) %>%
  #
  #save_kable("alpha-kable__________.png") # this line saves as .png in Reports/

all.bdiv.results
```
