==> R_scripts/05c_Plots_Phyla.Rmd <==
---
title: "R Notebook"
output: html_notebook
---

```{r}
rm(list=ls())
```

```{r}
#install.packages("sjPlot")
library(interactions)
library(jtools)
library(MuMIn)
library(lme4)
library(cowplot)
library(ggplot2)
library(sjPlot)
```

# Read in objects

```{r}
mot <- readRDS("../Data/05_melted.phyla.glom.rds")
```

```{r Split df by phyla}
proteobacteria.df <- mot[mot$Phylum=="Proteobacteria",]
firmicutes.df <- mot[mot$Phylum=="Firmicutes",]

```

Split df by phyla and subset to adults
```{r }
proteobacteria.adult.df <- mot[(mot$Phylum=="Proteobacteria" & mot$ageBinned=="adult"),]
firmicutes.adult.df <- mot[(mot$Phylum=="Firmicutes" & mot$ageBinned=="adult"),]
```

Read in dredged objects
```{r}
model.set.proteo <- readRDS("../Data/dredgeSaves/model.set.proteo")
model.set.firmicutes <- readRDS("../Data/dredgeSaves/model.set.firmicutes")
model.set.proteo.adult <- readRDS("../Data/dredgeSaves/model.set.proteo.adult")
model.set.firmicutes.adult <- readRDS(file = "../Data/dredgeSaves/model.set.firmicutes.adult")

```

Create objects for non-model averaged models
* seems to fit model when extracting it: glmer package and binomial response variable required
```{r}
binomial.response <- cbind(proteobacteria.df$Abundance/proteobacteria.df$NumberOfReads)
proteobacteria.model <- get.models(model.set.proteo, subset = 1)[[1]]

binomial.response <- cbind(firmicutes.df$Abundance/firmicutes.df$NumberOfReads)
firmicutes.model <- get.models(model.set.firmicutes, subset = 1)[[1]]

```

Adult top models
```{r}
binomial.response <- cbind(proteobacteria.adult.df$Abundance / proteobacteria.adult.df$NumberOfReads)
top.models.pro.adult <- get.models(model.set.proteo.adult, subset = delta<2)

binomial.response <- cbind(firmicutes.adult.df$Abundance / firmicutes.adult.df$NumberOfReads)
top.models.fir.adult <- get.models(model.set.firmicutes.adult, subset = delta<2)
```

##############################################################################################
##############################################################################################

Camille's templates (for categorical plots)
```{r}
# NameOfYourPlotWithInteractions<-cat_plot(NameOfYourModel,pred=YourPredictorFromInteraction,modx=YourOtherPredictorFromInteraction,x.label="NAMEX", y.label="NAMEY
# ",interval =TRUE,int.type = "confidence", partial.residuals = TRUE)
# 
# NameOfYourPlotWithOutInteractions<-effect_plot(NameOfYourModel,pred=YourPredictorOfInterest,x.label="Experimental Phase", y.label="Weighted
# Eigenvector Centrality",interval =TRUE,int.type = "confidence",partial.residuals = TRUE)
```


# All birds
## Age*habitat
Not a model averaged result, only 1 model in top set

Getting error in weighted mean default 
* Solved by eliminating "data =" argument

Why are some values of proteobacteria less than zero? Because "jitter" value was too low pushing points away from each other
```{r}
#check.df <- proteobacteria.model@frame

proteobacteria.model@frame$c.habitat <- ifelse(proteobacteria.model@frame$c.habitat<0, "Conifer", "Deciduous")

ageXhabitat.proteo.all<-cat_plot(proteobacteria.model, pred=ageBinned, modx=c.habitat, partial.residuals = TRUE, x.label="Age", y.label="Proteobacteria (relative abundance)", legend.main = "Habitat", pred.labels = c("Day 8", "Day 15", "Adult"), interval =TRUE, int.type = "confidence", jitter = 0.4, point.alpha = 0.2, point.size = 0.75)# data = proteobacteria.model@frame)

ageXhabitat.proteo.all
```

```{r}
#check.df <- proteobacteria.model@frame

firmicutes.model@frame$c.habitat <- ifelse(firmicutes.model@frame$c.habitat<0, "Conifer", "Deciduous")

ageXhabitat.firmic.all<-cat_plot(firmicutes.model, pred=ageBinned, modx=c.habitat, partial.residuals = TRUE, x.label="Age", y.label="Firmicutes (relative abundance)", legend.main = "Habitat", pred.labels = c("Day 8", "Day 15", "Adult"), interval =TRUE, int.type = "confidence", jitter = 0.4, point.alpha = 0.2, point.size = 0.75)# data = proteobacteria.model@frame)

ageXhabitat.firmic.all
```
###################################################
Re-read in these models because ifelse statement upstream, was upsetting the following plot codes
```{r}
binomial.response <- cbind(proteobacteria.df$Abundance/proteobacteria.df$NumberOfReads)
proteobacteria.model <- get.models(model.set.proteo, subset = 1)[[1]]

binomial.response <- cbind(firmicutes.df$Abundance/firmicutes.df$NumberOfReads)
firmicutes.model <- get.models(model.set.firmicutes, subset = 1)[[1]]


```
###################################################

## Distance
```{r}
distance.proteo.all <- effect_plot(proteobacteria.model, pred=z.DistanceToEdge,  partial.residuals = TRUE, x.label="Distance to edge", y.label="Proteobacteria (relative abundance)", interval =TRUE, int.type = "confidence", legend.main = "Age",  point.size = 0.75)

distance.proteo.all
```
###################################################
Re-read in these models because ifelse statement upstream, was upsetting the following plot codes
```{r}
binomial.response <- cbind(proteobacteria.df$Abundance/proteobacteria.df$NumberOfReads)
proteobacteria.model <- get.models(model.set.proteo, subset = 1)[[1]]

binomial.response <- cbind(firmicutes.df$Abundance/firmicutes.df$NumberOfReads)
firmicutes.model <- get.models(model.set.firmicutes, subset = 1)[[1]]


```
###################################################
## Age*broodSize

Should i back-transform or "unstandardise" brood size for the plot?
* can i just replace the standardised values in the model df with the actual values?

Error contrasts can be applied only to factors with 2 or more levels
* but above plot is identical and still works?
* can replace with proteobacteria model frame in hacky fix
* seems to be getting upset with some code around age*habitat plot as skipping age habitat plot code allows this to be plotted from fresh start
```{r}
#proteobacteria.model@frame$c.habitat <- ifelse(proteobacteria.model@frame$c.habitat<0, "conifer", "deciduous")

broodSizeXage.proteo.all<- interact_plot(proteobacteria.model, pred=z.broodSizeWhenSampled, modx=ageBinned, partial.residuals = TRUE, x.label="Brood Size", y.label="Proteobacteria (relative abundance)", interval =TRUE, int.type = "confidence", legend.main = "Age", modx.labels = c("Day 8", "Day 15", "Adult"), point.size = 0.75) 

broodSizeXage.proteo.all
```

```{r}
# firmicutes.model.HACK <- firmicutes.model
# firmicutes.model.HACK@frame <- proteobacteria.model@frame

broodSizeXage.firmic.all <- interact_plot(firmicutes.model, pred=z.broodSizeWhenSampled, modx=ageBinned, partial.residuals = TRUE, x.label="Brood Size", y.label="Firmicutes (relative abundance)", interval =TRUE, int.type = "confidence", legend.main = "Age", modx.labels = c("Day 8", "Day 15", "Adult"), point.size = 0.75) # data = proteobacteria.model@frame)

broodSizeXage.firmic.all
```
### age x brood size, standard mode fits
Using sjplot
```{r, eval = FALSE}
summary(firmicutes.model)

firmicutes.model@frame$ageBinned <- as.factor(firmicutes.model@frame$ageBinned)

str(firmicutes.model@frame$ageBinned)

plot_model(firmicutes.model, type = "pred", terms = c("z.broodSizeWhenSampled", "ageBinned"))

plot_model(firmicutes.model, type = "int")

#plot_model(firmicutes.model, type = "pred", terms = c("ageBinned", "z.broodSizeWhenSampled"))
```

Can plot with straight lines by specifying outcome.scale = link
Note: dropped partial residuals here aswell
```{r}
broodSizeXage.proteo.all.notPartial<- interact_plot(proteobacteria.model, pred=z.broodSizeWhenSampled, modx=ageBinned, partial.residuals = FALSE, x.label="Brood Size", y.label="Proteobacteria (relative abundance)", interval =TRUE, int.type = "confidence", legend.main = "Age", modx.labels = c("Day 8", "Day 15", "Adult"), point.size = 0.75, outcome.scale = "link") 

broodSizeXage.proteo.all.notPartial
```


## Age*Distance
```{r}
distanceXage.proteo.all <- interact_plot(proteobacteria.model, pred=z.DistanceToEdge, modx=ageBinned, partial.residuals = TRUE, x.label="Distance to edge", y.label="Proteobacteria (relative abundance)", interval =TRUE, int.type = "confidence", legend.main = "Age", modx.labels = c("Day 8", "Day 15", "Adult"), point.size = 0.75)

distanceXage.proteo.all
```

```{r}
distanceXage.firmic.all <- interact_plot(firmicutes.model, pred=z.DistanceToEdge, modx=ageBinned, partial.residuals = TRUE, x.label="Distance to edge", y.label="Firmicutes (relative abundance)", interval =TRUE, int.type = "confidence", legend.main = "Age", modx.labels = c("Day 8", "Day 15", "Adult"), point.size = 0.75) 

distanceXage.firmic.all
```

## age*layDate
```{r}
layDatexage.proteo.all <- interact_plot(proteobacteria.model, pred=z.layDateFirst, modx=ageBinned, partial.residuals = TRUE, x.label="Lay date", y.label="Proteobacteria (relative abundance)", interval =TRUE, int.type = "confidence", legend.main = "Age", modx.labels = c("Day 8", "Day 15", "Adult"), point.size = 0.75) 

layDatexage.proteo.all
```

```{r}
  
layDateXage.firmic.all <- interact_plot(firmicutes.model, pred=z.layDateFirst, modx=ageBinned, partial.residuals = TRUE, x.label="Lay date", y.label="Firmicutes (relative abundance)", interval =TRUE, int.type = "confidence", legend.main = "Age", modx.labels = c("Day 8", "Day 15", "Adult"), point.size = 0.75) 

layDateXage.firmic.all
```

# Adult birds
## Lay date
```{r}
#top.adult.proteo <- get.models(model.set.proteo.adult, subset = delta<2)
top.adult.proteo <- top.models.pro.adult[[1]]
#top.adult.proteo@frame$z.layDateFirst <- ifelse(top.adult.proteo@frame$c.ageDays>0,"Juvenile", "Mature")

layDate.proteo.adult<-effect_plot(top.adult.proteo, pred=z.layDateFirst, x.label="Lay date", y.label="Proteobacteria (relative abundance)", interval =TRUE, int.type = "confidence", partial.residuals = TRUE, jitter = 0.01, point.alpha = 0.5, point.size = 0.75) +ylim(0,1)

layDate.proteo.adult
```

Automatically making y axis between 0, 0.2. I will force to 0,1 to make more comparable
```{r}
#top.adult.proteo <- get.models(model.set.proteo.adult, subset = delta<2)
top.adult.firmic<- top.models.fir.adult[[2]] ## NOTE using 2nd model which contains lay date

layDate.firmic.adult<-effect_plot(top.adult.firmic, pred=z.layDateFirst, x.label="Lay date", y.label="Firmicutes (relative abundance)", interval =TRUE, int.type = "confidence", partial.residuals = TRUE, jitter = 0.01, point.alpha = 0.5, point.size = 0.75) + ylim(0,1)

layDate.firmic.adult
```


# Arrange plots

```{r}
theme_classic2 <- function(base_size = 12, base_family = ""){
  theme_bw(base_size = base_size, base_family = base_family) %+replace%
    theme(
      legend.position = "none",
      panel.border     = element_blank(),
      axis.line        = element_line(colour = "black"),
      panel.grid.major=element_line(colour="grey", size=0.5, 3),
      panel.grid.major.x = element_blank(),
      panel.grid.major.y = element_blank(),
      panel.grid.minor = element_blank(),
      panel.grid.minor.x = element_blank(),
      panel.grid.minor.y = element_blank(),
      strip.background = element_blank(),
      legend.key       = element_blank(),
      plot.title = element_text(hjust = 0.5, face = "bold") # centre and bold title
    )
}
```

Adding this theme to get rid of axis lines in main phyla plot
```{r}
theme_classic3 <- function(base_size = 12, base_family = ""){
  theme_bw(base_size = base_size, base_family = base_family) %+replace%
    theme(
      legend.position = "none",
      panel.border     = element_blank(),
      axis.line        = element_line(colour = "white"),
      panel.grid.major=element_line(colour="grey", size=0.5, 3),
      panel.grid.major.x = element_blank(),
      panel.grid.major.y = element_blank(),
      panel.grid.minor = element_blank(),
      panel.grid.minor.x = element_blank(),
      panel.grid.minor.y = element_blank(),
      strip.background = element_blank(),
      legend.key       = element_blank(),
      plot.title = element_text(hjust = 0.5, face = "bold") # centre and bold title
    )
}
```

```{r}
p3 <- cowplot::plot_grid(
  ageXhabitat.proteo.all+theme_classic2()+ylab(NULL)+labs(title="Proteobacteria"),
  ageXhabitat.firmic.all+theme_classic2()+ylab(NULL)+labs(title="Firmicutes"),
  ncol=2, labels = c("a", "b"))

p4 <- cowplot::plot_grid(broodSizeXage.proteo.all+theme_classic2()+ylab(NULL), broodSizeXage.firmic.all+theme_classic2()+ylab(NULL), distanceXage.proteo.all+theme_classic2()+ylab(NULL), distanceXage.firmic.all+theme_classic2()+ylab(NULL), layDatexage.proteo.all+theme_classic2()+ylab(NULL), layDateXage.firmic.all+theme_classic2()+ylab(NULL), ncol=2, labels = c("c", "d", "e", "f", "g", "h"), label_x = 0, label_y = 1.2)

# get common legend of all interaction plots
p3.legend <- get_legend(ageXhabitat.proteo.all)
p4.legend <- get_legend(broodSizeXage.proteo.all)

p3.grid <- cowplot::plot_grid(p3, p3.legend, rel_widths = c(3, .5))
p4.grid <- cowplot::plot_grid(p4, p4.legend, rel_widths = c(3, .5))

p5 <- cowplot::plot_grid(p3.grid, p4.grid, nrow = 2,  rel_heights = c(1,3)) + theme_classic3()
p5
```

```{r}
ggsave("Fig4_RA_allBirds_combined.pdf", plot = p5, device = "pdf", path = "../R_figures", width = 20, height = 20, units = "cm", dpi = 200)
```

```{r}
p1 <- cowplot::plot_grid(layDate.proteo.adult+theme_classic2()+ylab("Relative abundance")+labs(title="Proteobacteria"), layDate.firmic.adult+theme_classic2()+ylab(NULL)+labs(title="Firmicutes"), labels = c("A", "B")) 

p1
```

```{r}
ggsave("Fig5_RA_adults_layDate.tiff", plot = p1, device = "tiff", path = "../R_figures", width = 16, height = 8, units = "cm", dpi = 150)
```

# Plot model averaged models

Error: $ operator is invalid for atomic vectorsxs
```{r}
# model.avg.proteobacteria.adult <- model.avg(top.models.pro.adult)
# 
# interact_plot(model.avg.proteobacteria.adult, pred=z.layDateFirst, modx=c.Sex, interval =TRUE, int.type = "confidence", x.label="Lay date", y.label="Firmicutes (relative abundance)", legend.main = "Sex" , data = top.adult.proteo@frame)
```

# Plots for HNB Oxford

```{r}
ageXhabitat.only <- cowplot::plot_grid(
  ageXhabitat.proteo.all+theme_classic2()+ylab(NULL)+labs(title="Proteobacteria"),
  ageXhabitat.firmic.all+theme_classic2()+ylab(NULL)+labs(title="Firmicutes"),
  ncol=2)

#ageXhabitat.only.grid <- cowplot::plot_grid(ageXhabitat.only, p3.legend, rel_widths = c(3, .5))
ageXhabitat.only.grid <- cowplot::plot_grid(p3.legend, ageXhabitat.only, nrow = 2, rel_heights = c(1,5))

ageXhabitat.only.grid

ggsave("hnb_ageXhabitat_RA.tiff", plot = ageXhabitat.only.grid, device = "tiff", path = "../R_figures/HNB", width = 16, height = 12, units = "cm", dpi = 150)
```


```{r}
distancexage.only <- cowplot::plot_grid(
  distanceXage.proteo.all+theme_classic2()+ylab(NULL)+labs(title="Proteobacteria"),
  distanceXage.firmic.all+theme_classic2()+ylab(NULL)+labs(title="Firmicutes"),
  ncol=2)

#distancexage.only.grid <- cowplot::plot_grid(distancexage.only, p4.legend, rel_widths = c(3, .5))
#distancexage.only.grid <- cowplot::plot_grid(distancexage.only, p4.legend, nrow = 2, rel_heights = c(3, 1))

distancexage.only.grid <- cowplot::plot_grid(p4.legend, distancexage.only, nrow = 2, rel_heights = c(1,4))

distancexage.only.grid

ggsave("hnb_distanceXage_RA.tiff", plot = distancexage.only.grid, device = "tiff", path = "../R_figures/HNB", width = 22, height = 14, units = "cm", dpi = 150)
```

```{r}
broodSizeXage.only <- cowplot::plot_grid(
  broodSizeXage.proteo.all+theme_classic2()+ylab(NULL)+labs(title="Proteobacteria"),
  broodSizeXage.firmic.all+theme_classic2()+ylab(NULL)+labs(title="Firmicutes"),
  ncol=2)

broodSizeXage.only.grid <- cowplot::plot_grid(p4.legend, broodSizeXage.only, nrow = 2, rel_heights = c(1, 4))

broodSizeXage.only.grid

ggsave("hnb_broodXage_RA.tiff", plot = broodSizeXage.only.grid, device = "tiff", path = "../R_figures/HNB", width = 20, height = 14, units = "cm", dpi = 150)
```

```{r}
layDateXage.only <- cowplot::plot_grid(
  layDatexage.proteo.all+theme_classic2()+ylab(NULL)+labs(title="Proteobacteria"),
  layDateXage.firmic.all+theme_classic2()+ylab(NULL)+labs(title="Firmicutes"),
  ncol=2)

#layDateXage.only.grid <- cowplot::plot_grid(layDateXage.only, p4.legend, rel_widths = c(3, .5))
layDateXage.only.grid <- cowplot::plot_grid(p4.legend, layDateXage.only, nrow = 2, rel_heights = c(1, 4))

layDateXage.only.grid

ggsave("hnb_laydateXage_RA.tiff", plot = layDateXage.only.grid, device = "tiff", path = "../R_figures/HNB", width = 20, height = 14, units = "cm", dpi = 150)
```
